Implement VS Extension Settings Persistence (#508)
* Save Settings
---------
