+++ DevSkim-DotNet/Microsoft.DevSkim.VisualStudio/Microsoft.DevSkim.VisualStudio.csproj
@@ -52,7 +52,7 @@
     <Compile Include="DevSkimFixMessageTarget.cs" />
     <Compile Include="DevSkimLanguageClient.cs" />
     <Compile Include="DevSkimSuggestedAction.cs" />
-    <Compile Include="Options\FindingOptionsPage.cs">
+    <Compile Include="Options\GeneralOptionsPage.cs">
       <SubType>Component</SubType>
     </Compile>
     <Compile Include="ProcessTracker\IProcessTracker.cs" />
@@ -63,26 +63,13 @@
     <Compile Include="ProcessTracker\WindowEventArgs.cs" />
     <Compile Include="ProcessTracker\WindowEventsListener.cs" />
     <Compile Include="ProcessTracker\WindowsSystemEvents.cs" />
-    <Compile Include="Options\IgnoreOptionsPage.cs">
-      <SubType>Component</SubType>
-    </Compile>
-    <Compile Include="Options\GuidanceOptionsPage.cs">
-      <SubType>Component</SubType>
-    </Compile>
-    <Compile Include="Options\AllOptionsPage.cs">
-      <SubType>Component</SubType>
-    </Compile>
-    <Compile Include="Options\SuppressionOptionsPage.cs">
-      <SubType>Component</SubType>
-    </Compile>
-    <Compile Include="Options\RuleOptionsPage.cs">
-      <SubType>Component</SubType>
-    </Compile>
-    <Compile Include="Options\TriggerOptionsPage.cs">
-      <SubType>Component</SubType>
-    </Compile>
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Options\OptionsPackage.cs" />
+    <Compile Include="Properties\Resources.Designer.cs">
+      <AutoGen>True</AutoGen>
+      <DesignTime>True</DesignTime>
+      <DependentUpon>Resources.resx</DependentUpon>
+    </Compile>
     <Compile Include="SettingsChangedNotifier.cs" />
     <Compile Include="StaticData.cs" />
     <Compile Include="SuggestedActionsSourceProvider.cs" />
@@ -152,6 +139,12 @@
       <IncludeInVSIX>true</IncludeInVSIX>
       <VSIXSubPath>Content</VSIXSubPath>
     </Content>
+  </ItemGroup>
+  <ItemGroup>
+    <EmbeddedResource Include="Properties\Resources.resx">
+      <Generator>ResXFileCodeGenerator</Generator>
+      <LastGenOutput>Resources.Designer.cs</LastGenOutput>
+    </EmbeddedResource>
   </ItemGroup>
   <!-- Publish the language server binaries for Release -->
   <Target Name="BuildLanguageServer" BeforeTargets="IncludeLanguageServer" Condition=" '$(Configuration)' == 'Release' ">
