--- DevSkim-DotNet/Microsoft.DevSkim.VisualStudio/Options/AllOptionsPage.cs
@@ -1,13 +1,0 @@
-﻿namespace Microsoft.DevSkim.VisualStudio
-{
-    using Microsoft.VisualStudio.Shell;
-    using System;
-    using System.Collections.Generic;
-    using System.ComponentModel;
-    using System.Windows.Media.Media3D;
-
-    public class AllOptionsPage : DialogPage
-    {
-        
-    }
-}
